requirements.txt
pandas
numpy
scikit-learn
xgboost
tensorflow
statsmodels
matplotlib
yfinance
fredapi
tweepy
textblob
shap
python-dotenv
backtrader
